Interpolate with constraints
Works on a dictionary of constraints. Just leave values blank if you don't want to impose any constraints. Otherwise, put in an entry with the variable name as the key and a tuple of the bottom range and top range for the value.

If you want to constrain a variable to a certain value, put the same value for as the first and second entry of the tuple.

Example: {"orifice_size":(100,100)} leaves the orifice size as 100